ADD: user permission & ownership of images

+++ api/photos/delete.php
@@ -1,9 +1,17 @@
 <?php
 // Includes
-include_once('../../' . dirname('.') . '/lib/db.php');
-include_once('../../' . dirname('.') . '/models/photo.php');
+include_once('/app/models/photo.php');
+include_once('/app/includes/header.inc');
 
 $id = $_GET['id'];
 $PhotoClass = new Photo();
-$deleteMessage = $PhotoClass->deletePhoto($id);
-header("Location: http://localhost/photos?message=" . $deleteMessage, true, 302);
+
+$photo = $PhotoClass->singlePhoto($id);
+
+if (!($session->get('user.id') == $photo['user_id'])) {
+  header("Location: http://localhost/photos?message=notAllowed", true, 302);
+
+} else {
+  $deleteMessage = $PhotoClass->deletePhoto($id);
+  header("Location: http://localhost/photos?message=" . $deleteMessage, true, 302);
+}
